+++ library/src/main/java/com/bumptech/glide/load/resource/bitmap/UnitBitmapDecoder.java
@@ -5,16 +5,15 @@ import android.support.annotation.Nullable;
 import com.bumptech.glide.load.Options;
 import com.bumptech.glide.load.ResourceDecoder;
 import com.bumptech.glide.load.engine.Resource;
-import com.bumptech.glide.load.engine.bitmap_recycle.BitmapPool;
+import com.bumptech.glide.load.engine.bitmap_recycle.BitmapPoolAdapter;
 import java.io.IOException;
 
+/**
+ * Passes through a (hopefully) non-owned {@link Bitmap} as a {@link Bitmap} based {@link Resource}
+ * so that the given {@link Bitmap} is not recycled.
+ */
 public final class UnitBitmapDecoder implements ResourceDecoder<Bitmap, Bitmap> {
-
-  private final BitmapPool bitmapPool;
-
-  public UnitBitmapDecoder(BitmapPool bitmapPool) {
-    this.bitmapPool = bitmapPool;
-  }
+  private static final BitmapPoolAdapter BITMAP_POOL = new BitmapPoolAdapter();
 
   @Override
   public boolean handles(Bitmap source, Options options) throws IOException {
@@ -25,6 +24,6 @@ public final class UnitBitmapDecoder implements ResourceDecoder<Bitmap, Bitmap> 
   @Override
   public Resource<Bitmap> decode(Bitmap source, int width, int height, Options options)
       throws IOException {
-    return new BitmapResource(source, bitmapPool);
+    return new BitmapResource(source, BITMAP_POOL);
   }
 }
